flask-server/requirements.txt
Flask==2.3.3
flask-pymongo==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Cors==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
